Web_App_Firewall/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
redis==4.6.0
gunicorn==21.2.0
pytest==7.4.2
black==23.7.0
flake8==6.0.0
docker==6.1.3 